working example of content locale switching
