+++ app/content/index.js
@@ -1,0 +1,9 @@
+import en from './en'
+import es from './es'
+
+const content = {
+  en: en,
+  es: es
+}
+
+export default content
